# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: billing
    description: Accounts, entitlement and membership subscription
  - name: ask
    description: Grounded streaming assistant about the scientist's work
  - name: feedback
    description: Bug reports and feature requests
  - name: ship
    description: The signed-in account's saved cosmonaut ship
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /me/entitlement:
    get:
      operationId: getEntitlement
      tags: [billing]
      summary: Current account entitlement
      description: >
        Returns whether the signed-in account is an active member and which
        researchers it has unlocked. Requires an authenticated session; the
        galaxy default researcher never calls this since it is always free.
      responses:
        "200":
          description: Entitlement state for the signed-in account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entitlement"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me/unlock:
    post:
      operationId: unlockResearcher
      tags: [billing]
      summary: Unlock a researcher for the signed-in member
      description: >
        Unlocks a researcher's galaxy for an active member. The first few
        researchers are included in the base membership; each one beyond that
        adds a recurring +$1/year add-on, charged immediately with proration.
        Idempotent — re-unlocking an already-unlocked researcher never charges.
        Returns the updated entitlement.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnlockRequest"
      responses:
        "200":
          description: Updated entitlement state after the unlock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entitlement"
        "400":
          description: Invalid researcher id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "402":
          description: Not a member, or the add-on charge could not be collected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /billing/checkout:
    post:
      operationId: createCheckout
      tags: [billing]
      summary: Start the membership subscription checkout
      description: >
        Creates a Stripe Checkout session (mode=subscription) for the $7/year
        full-access membership and returns its hosted URL. If the account is
        already entitled, returns alreadyEntitled=true and no URL.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout session created (or account already entitled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /billing/confirm:
    post:
      operationId: confirmCheckout
      tags: [billing]
      summary: Confirm a completed checkout session
      description: >
        Verifies a returned Stripe Checkout session directly against Stripe and,
        when paid and owned by the signed-in account, marks the account entitled.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmRequest"
      responses:
        "200":
          description: Updated entitlement state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entitlement"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me/ship:
    get:
      operationId: getShip
      tags: [ship]
      summary: The signed-in account's saved ship
      description: >
        Returns the seed of the ship the account has saved, or null if none has
        been saved yet. The ship look is derived deterministically from this
        seed on the client.
      responses:
        "200":
          description: The account's saved ship seed (or null)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipState"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: saveShip
      tags: [ship]
      summary: Save the account's ship
      description: >
        Persists the account's chosen ship seed so it follows them across
        devices and is broadcast to other cosmonauts. The seed is sanitized to a
        short alphanumeric string server-side.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveShipRequest"
      responses:
        "200":
          description: The saved ship seed (after sanitization)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipState"
        "400":
          description: Invalid seed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me/ship/claim:
    post:
      operationId: claimSkin
      tags: [ship]
      summary: Claim or purchase a premium ship type
      description: >
        Claims a premium ship type for the signed-in account. Active members get
        a fixed number of types included for free; once those are used (or for
        non-members) the account pays $1 one-time. Returns alreadyOwned when the
        account already has it, granted when a free member slot was applied, or a
        hosted Stripe Checkout URL when payment is required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimSkinRequest"
      responses:
        "200":
          description: Claim outcome (owned, granted free, or checkout URL)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkinCheckout"
        "400":
          description: Invalid or unknown ship type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me/ship/confirm:
    post:
      operationId: confirmSkin
      tags: [ship]
      summary: Confirm a completed ship-type checkout
      description: >
        Verifies a returned Stripe Checkout session directly against Stripe and,
        when paid and owned by the signed-in account, grants the ship type.
        Returns the refreshed ship state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmRequest"
      responses:
        "200":
          description: Refreshed ship state after the grant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipState"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me/referral:
    get:
      operationId: getReferral
      tags: [referral]
      summary: The signed-in account's referral link info
      description: >
        Returns the signed-in account's stable referral code and how many
        accounts have signed up through their link. The code is generated on
        first call and is permanent.
      responses:
        "200":
          description: The account's referral code and signup count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralInfo"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me/referral/claim:
    post:
      operationId: claimReferral
      tags: [referral]
      summary: Attribute the signed-in (newly created) account to a referrer
      description: >
        Called right after sign-up with a referral code captured from a `?ref=`
        link. Attributes this account to the referrer — but only for genuinely
        new accounts and only once (the attribution is permanent). Safe to call
        when ineligible: it simply returns the account's own referral info
        unchanged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimReferralRequest"
      responses:
        "200":
          description: The account's own referral info after the best-effort claim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralInfo"
        "400":
          description: Missing referral code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /ask/chat:
    post:
      operationId: chatAsk
      tags: [ask]
      summary: Ask the grounded streaming assistant a question
      description: >
        A grounded "router + writer": the model classifies the turn (data /
        explain / chat / feedback), emits a structured action on the first
        streamed line, then streams a short reasoning trace and the answer. The
        model never computes filtered counts or lists — for "data" turns it only
        fills a query spec the browser runs deterministically over the
        locally-baked data; it may only repeat the exact summary stats it is
        given. Only the question, the data shape, the domain names, and the
        public corpus summary are sent; no actual paper data leaves the browser.
        The response is a Server-Sent-Events stream, read manually on the client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskRequest"
      responses:
        "200":
          description: An SSE stream of action, reasoning and answer events
          content:
            text/event-stream: {}
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: The assistant model is unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /feedback/issue:
    post:
      operationId: reportFeedback
      tags: [feedback]
      summary: File a bug report or feature request as a private EXO work item
      description: >
        Creates a private work item through EXO's signed intake boundary from a
        visitor's message. Used by the Ask panel's "report a bug / request a
        feature" affordance.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "201":
          description: The EXO work item was created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: EXO intake is unreachable or work-item creation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
        deployment:
          type: object
          properties:
            commit:
              type: string
            environment:
              type: string
          required:
            - commit
            - environment
      required:
        - status
        - deployment
    Entitlement:
      type: object
      properties:
        entitled:
          type: boolean
          description: Whether the account is an active full-access member.
        email:
          type: string
          nullable: true
          description: The account's primary email, when known.
        includedSlots:
          type: integer
          description: How many researcher unlocks the base membership includes.
        unlocked:
          type: array
          description: OpenAlex author ids this account has unlocked, oldest first.
          items:
            type: string
      required:
        - entitled
        - includedSlots
        - unlocked
    UnlockRequest:
      type: object
      properties:
        author:
          type: string
          description: OpenAlex author id of the researcher to unlock.
      required:
        - author
    CheckoutSession:
      type: object
      properties:
        url:
          type: string
          nullable: true
          description: Hosted Stripe Checkout URL to redirect to, when a session was created.
        alreadyEntitled:
          type: boolean
          description: True when the account is already an active member; no checkout was created.
      required:
        - alreadyEntitled
    CheckoutRequest:
      type: object
      properties:
        author:
          type: string
          nullable: true
          description: >
            OpenAlex author id of the scientist currently being explored, so the
            post-payment redirect returns to that galaxy instead of the default.
    ConfirmRequest:
      type: object
      properties:
        sessionId:
          type: string
          description: The Stripe Checkout session id returned on the success redirect.
      required:
        - sessionId
    ShipState:
      type: object
      properties:
        shipSeed:
          type: string
          nullable: true
          description: The account's saved ship seed, or null if none is saved.
        shipType:
          type: string
          description: The equipped ship type id (defaults to "scout").
        ownedTypes:
          type: array
          description: >
            Ship type ids the account can equip — always includes the free
            "scout", plus every premium type it has claimed or purchased.
          items:
            type: string
        entitled:
          type: boolean
          description: Whether the account is an active member (free-slot eligible).
        includedSkinSlots:
          type: integer
          description: How many premium ship types the membership includes for free.
        freeSlotsRemaining:
          type: integer
          description: >
            Included premium-type slots not yet used (members only; 0 otherwise).
      required:
        - shipSeed
        - shipType
        - ownedTypes
        - entitled
        - includedSkinSlots
        - freeSlotsRemaining
    SaveShipRequest:
      type: object
      properties:
        seed:
          type: string
          description: The ship seed to save (sanitized to short alphanumeric server-side).
        type:
          type: string
          nullable: true
          description: >
            The ship type id to equip. Must be owned by the account (or the free
            "scout"); defaults to "scout" when omitted.
      required:
        - seed
    ClaimSkinRequest:
      type: object
      properties:
        type:
          type: string
          description: The premium ship type id to claim or purchase.
        author:
          type: string
          nullable: true
          description: >
            OpenAlex author id of the scientist currently being explored, so the
            post-payment redirect returns to that galaxy instead of the default.
      required:
        - type
    SkinCheckout:
      type: object
      properties:
        alreadyOwned:
          type: boolean
          description: True when the account already owns this type; no action taken.
        granted:
          type: boolean
          description: True when a free member-slot claim was applied (no payment).
        url:
          type: string
          nullable: true
          description: Hosted Stripe Checkout URL to redirect to, when payment is required.
      required:
        - alreadyOwned
        - granted
    AskField:
      type: object
      description: A single field of the data shape (name + type), no actual data.
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
      required:
        - name
        - type
    AskMessage:
      type: object
      description: One turn of the ephemeral, in-browser conversation.
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content
    AskSummary:
      type: object
      description: >
        The public, headline corpus summary the model may quote verbatim. These
        are the ONLY numbers the model is allowed to state; any filtered/derived
        count must route to a "data" action for deterministic in-browser compute.
      properties:
        authorName:
          type: string
        institution:
          type: string
          nullable: true
        totalPapers:
          type: integer
        totalCitations:
          type: integer
        hIndex:
          type: integer
          nullable: true
        i10Index:
          type: integer
          nullable: true
        uniqueCoAuthors:
          type: integer
        firstYear:
          type: integer
        lastYear:
          type: integer
        avgCitations:
          type: integer
        topDomains:
          type: array
          description: The largest research domains, biggest first.
          items:
            type: string
        mostCitedTitle:
          type: string
          nullable: true
        mostCitedCount:
          type: integer
          nullable: true
      required:
        - authorName
        - totalPapers
        - totalCitations
        - uniqueCoAuthors
        - firstYear
        - lastYear
        - avgCitations
        - topDomains
    AskRequest:
      type: object
      properties:
        messages:
          type: array
          description: The ephemeral conversation so far (oldest first).
          items:
            $ref: "#/components/schemas/AskMessage"
        fields:
          type: array
          description: The shape of a paper record (field names and types only).
          items:
            $ref: "#/components/schemas/AskField"
        domains:
          type: array
          description: The research-domain (category) names present in this galaxy.
          items:
            type: string
        summary:
          $ref: "#/components/schemas/AskSummary"
      required:
        - messages
        - summary
    FeedbackRequest:
      type: object
      properties:
        kind:
          type: string
          enum: [bug, feature]
          description: Whether this is a bug report or a feature request.
        message:
          type: string
          description: The visitor's message.
      required:
        - kind
        - message
    FeedbackResult:
      type: object
      properties:
        trackingId:
          type: string
          description: Opaque EXO tracking id retained by the application.
        identifier:
          type: string
          description: Human-readable EXO work-item identifier.
      required:
        - trackingId
        - identifier
    ReferralInfo:
      type: object
      properties:
        code:
          type: string
          description: The account's stable referral code, used as `?ref=<code>`.
        referredCount:
          type: integer
          description: How many accounts have signed up through this account's link.
      required:
        - code
        - referredCount
    ClaimReferralRequest:
      type: object
      properties:
        code:
          type: string
          description: The referral code captured from the `?ref=` link at sign-up.
      required:
        - code
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error