# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HealthStatusDeployment } from './healthStatusDeployment';

export interface HealthStatus {
  status: string;
  deployment: HealthStatusDeployment;
}